refactor: change naming (app -> server) + small fixes
