+++ src/index.ts
@@ -1,8 +1,8 @@
-import { server } from './app';
+import { server } from './server';
 import { getPort } from './utils';
 
 const PORT = getPort();
 
-server.listen(PORT, () => {
-  console.log(`server started on port: ${PORT}`);
+server.listen(PORT, (): void => {
+  console.log(`Server started on port: ${PORT}`);
 });
