+++ src/utils/getPort.ts
@@ -5,7 +5,10 @@ export const getPort = (): number => {
   const PORT: string | undefined = process.env.PORT;
 
   if (!PORT) {
-    console.log(`.env doesn't have PORT constant. PORT switched to deafult value (4000)`);
+    console.log(`
+    .env doesn't have PORT constant.
+    PORT switched to deafult value (4000).
+    `);
     return DEFAULT_PORT;
   }
